修改优化

+++ app/Http/Controllers/Admin/ArticleController.php
@@ -56,8 +56,21 @@ class ArticleController extends AdminBaseController
     {
         $result = Article::find($id);
         if ($result) {
-            $result->category_name = $result->categorys?$result->categorys->name:"";
+//            $result->category_name = $result->categorys?$result->categorys->name:"";
             $article = $result->toArray();
+            if ($result->has('tags')->count()){
+                foreach ($result->tags as $item){
+                    $article['tags'][] = $item->name;
+                }
+//                $tags = Tag::whereIn('id',$result->tags);
+//                $article['tags'] = array_map(function(){
+//
+//                },$result->tags);
+            }else{
+                $article['tags'] = [];
+            }
+
+
             return $this->sendSuccess($article);
         }
             return $this->sendFail("没有找到该文章信息！");
@@ -66,15 +79,33 @@ class ArticleController extends AdminBaseController
     public function update($id=0,Request $request)
     {
         $data = $request->all()['form_data'];
-        $result = Article::where("article_id", $id)->update($data);
+        $tags = [];
+        if ($data['tags']){
+            foreach ($data['tags'] as $item){
+                $tag = Tag::firstOrCreate(['name' => $item]);
+                $tags[] = $tag->id;
+            }
+        }
+        unset($data['tags']);
+        $article = Article::find($id);
+        $result = $article->update($data);
+        if ($tags){
+            $article->tags()->sync($tags);
+        }
         if (!$result) {
             return $this->sendFail("修改失败！");
         }
+        return $this->sendSuccess();
     }
     public function destroy($id=0)
     {
+        if (!$id){
+            return $this->sendFail("缺少参数");
+        }
+        $article = Article::find($id);
+        $tag_result = $article->tags()->detach();
+        $common_result = $article->comments()->delete();
         $result=Article::destroy($id);
-
         if ($result){
             return $this->sendSuccess("","删除文章成功！");
         }
